+++ bot/utility/strings.py
@@ -2,5 +2,17 @@
 
 DM_WELCOME_MESSAGE = '''Hey there {}! Welcome to the 87th Regiment of Foot!\n
 If you are interested in joining us, head over to #role-registry-office and fill out the pinned form. If you are looking to transfer from another regiment, head over to #transfers-office.\n
-If you are from another regiment and need roles, let us know!\n\n 
-We look forward to seeing you on the battlefield! ⚔️'''
+If you are from another regiment and need roles, let us know!\n 
+We look forward to seeing you on the battlefield! ⚔️\n\n
+
+*I am a bot, this message was sent automatically :)*'''
+
+ENLIST_SUCCESS_MESSAGE = '''Hey there <@{}>! Welcome to the 87th Regiment of Foot!\n
+You have been enlisted successfully. There are just a few housekeeping tasks to tidy up:
+
+- Make sure to join the in-game registry. Search for '87th' and select the one led by GodEmperorBrolly.
+- Ensure that in your Holdfast options, 'Use Regiment Registry Name' is enabled. This means you'll get our tags in-game
+- Join our steam group, the link for which is in #regiment-documentation
+
+We'll see you on the battlefield! If you have any questions, please don't hesitate to ask! ⚔️
+'''